requirements.txt
insightface
faiss-cpu
opencv-python
numpy
pandas
albumentations
torch
torchvision
scikit-learn
fastapi
uvicorn[standard]
onnxruntime
python-multipart
sqlalchemy
psycopg2-binary